move load_corpus to python module
