+++ backend/conftest.py
@@ -5,7 +5,7 @@ import requests
 from allauth.account.models import EmailAddress
 
 from ianalyzer.elasticsearch import elasticsearch
-from addcorpus.load_corpus import load_corpus_definition
+from addcorpus.python_corpora.load_corpus import load_corpus_definition
 from addcorpus.save_corpus import load_and_save_all_corpora
 from es import es_index as index
