+++ backend/corpora/dutchannualreports/dutchannualreports.py
@@ -11,7 +11,7 @@ from addcorpus.extract import XML, Metadata, Combined
 from addcorpus.filters import MultipleChoiceFilter, RangeFilter
 from addcorpus.corpus import XMLCorpusDefinition, FieldDefinition
 from media.image_processing import get_pdf_info, retrieve_pdf, pdf_pages, build_partial_pdf
-from addcorpus.load_corpus import corpus_dir
+from addcorpus.python_corpora.load_corpus import corpus_dir
 from addcorpus.es_mappings import keyword_mapping, main_content_mapping
 from addcorpus.es_settings import es_settings
