+++ backend/corpora/dutchnewspapers/dutchnewspapers_public.py
@@ -13,7 +13,7 @@ from django.conf import settings
 from addcorpus.corpus import XMLCorpusDefinition, FieldDefinition, consolidate_start_end_years
 from addcorpus import filters
 from addcorpus.extract import Metadata, XML
-from addcorpus.load_corpus import corpus_dir
+from addcorpus.python_corpora.load_corpus import corpus_dir
 
 from corpora.utils.constants import document_context
 from addcorpus.es_mappings import keyword_mapping, main_content_mapping
