+++ backend/corpora/troonredes/troonredes.py
@@ -6,18 +6,14 @@ locations.
 import logging
 logger = logging.getLogger(__name__)
 import os
-from os.path import join, isfile, splitext, isfile
-from datetime import datetime, timedelta
-import re
-import random
-from pprint import pprint
+from os.path import join, splitext
+from datetime import datetime
 
 from django.conf import settings
 
 from addcorpus import extract
 from addcorpus import filters
-from addcorpus.corpus import XMLCorpusDefinition, FieldDefinition, until, after, string_contains
-from addcorpus.load_corpus import corpus_dir
+from addcorpus.corpus import XMLCorpusDefinition, FieldDefinition
 
 from addcorpus.es_mappings import keyword_mapping, main_content_mapping
 from addcorpus.es_settings import es_settings
