+++ backend/es/conftest.py
@@ -3,7 +3,7 @@ from time import sleep
 from django.contrib.auth.models import Group
 
 from addcorpus.conftest import basic_corpus
-from addcorpus.load_corpus import load_corpus_definition
+from addcorpus.python_corpora.load_corpus import load_corpus_definition
 from addcorpus.models import Corpus
 from es import es_index
