+++ backend/es/management/commands/alias.py
@@ -1,6 +1,6 @@
 from django.core.management import BaseCommand
 
-from addcorpus.load_corpus import load_corpus_definition
+from addcorpus.python_corpora.load_corpus import load_corpus_definition
 from es.es_alias import alias
 
 class Command(BaseCommand):
