+++ backend/visualization/wordcloud.py
@@ -1,7 +1,7 @@
 from collections import Counter
 from sklearn.feature_extraction.text import CountVectorizer
 
-from addcorpus.load_corpus import load_corpus_definition
+from addcorpus.python_corpora.load_corpus import load_corpus_definition
 from addcorpus.es_settings import get_stopwords_from_settings
 from es import download as download
 
@@ -21,7 +21,7 @@ def make_wordcloud_data(documents, field, corpus):
         content = document['_source'][field]
         if content and content != '':
             texts.append(content)
-            
+
     stopwords = field_stopwords(corpus, field)
     cv = CountVectorizer(max_features=100, max_df=0.7, token_pattern=r'(?u)\b[^0-9\s]{3,30}\b', stop_words=stopwords)
     cvtexts = cv.fit_transform(texts)
